commitlog: move serialization/deserialization to `commitlog_entry_io`
This change refactors commitlog_entry by moving `commitlog_entry_reader` and `commitlog_entry_writer`
to separate files: `commitlog_entry_io.hh` and `commitlog_entry_io.cc`.

This is needed for further changes in the scope of strong consistency for shared commit log–based
persistence for Raft groups.
