+++ configure.py
@@ -1119,7 +1119,7 @@ scylla_core = (['message/messaging_service.cc',
                 'db/batchlog_manager.cc',
                 'db/corrupt_data_handler.cc',
                 'db/commitlog/commitlog.cc',
-                'db/commitlog/commitlog_entry.cc',
+                'db/commitlog/commitlog_entry_io.cc',
                 'db/commitlog/commitlog_replayer.cc',
                 'db/config.cc',
                 'db/consistency_level.cc',
